backend/requirements.txt
flask
openai
transformers
torch
pdfplumber
python-dotenv
flask-cors
requests